+++ src/main/kotlin/org/wfanet/panelmatch/client/common/EventCompressorTrainer.kt
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.wfanet.panelmatch.client.eventpreprocessing
+package org.wfanet.panelmatch.client.common
 
 import com.google.protobuf.ByteString
 import java.io.Serializable
