+++ src/main/kotlin/org/wfanet/panelmatch/client/common/testing/BUILD.bazel
@@ -14,9 +14,11 @@ kt_jvm_library(
     deps = [
         "//imports/java/org/apache/beam:core",
         "//imports/java/org/apache/beam/extensions:protobuf",
+        "//src/main/kotlin/org/wfanet/panelmatch/client/common",
         "//src/main/kotlin/org/wfanet/panelmatch/common",
         "//src/main/kotlin/org/wfanet/panelmatch/common/beam",
         "//src/main/kotlin/org/wfanet/panelmatch/common/beam/testing",
+        "//src/main/kotlin/org/wfanet/panelmatch/common/compression/testing",
         "@wfa_common_jvm//imports/java/com/google/protobuf",
         "@wfa_common_jvm//imports/java/org/junit",
     ],
