+++ src/main/kotlin/org/wfanet/panelmatch/client/common/testing/FakeEventCompressorTrainer.kt
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.wfanet.panelmatch.client.eventpreprocessing.testing
+package org.wfanet.panelmatch.client.common.testing
 
 import com.google.protobuf.ByteString
-import org.wfanet.panelmatch.client.eventpreprocessing.EventCompressorTrainer
-import org.wfanet.panelmatch.client.eventpreprocessing.EventCompressorTrainer.TrainedEventCompressor
+import org.wfanet.panelmatch.client.common.EventCompressorTrainer
+import org.wfanet.panelmatch.client.common.EventCompressorTrainer.TrainedEventCompressor
 import org.wfanet.panelmatch.common.compression.testing.FakeCompressor
 import org.wfanet.panelmatch.common.toByteString
