+++ src/main/kotlin/org/wfanet/panelmatch/client/eventpreprocessing/BUILD.bazel
@@ -37,7 +37,6 @@ kt_jvm_library(
         "//src/main/kotlin/org/wfanet/panelmatch/client/logger",
         "//src/main/kotlin/org/wfanet/panelmatch/common",
         "//src/main/kotlin/org/wfanet/panelmatch/common/beam",
-        "//src/main/kotlin/org/wfanet/panelmatch/common/compression",
         "//src/main/proto/wfa/panelmatch/client/eventpreprocessing:combined_events_kt_jvm_proto",
         "//src/main/proto/wfa/panelmatch/client/eventpreprocessing:preprocess_events_kt_jvm_proto",
         "//src/main/swig/wfanet/panelmatch/client/eventpreprocessing:preprocess_events",
