+++ src/main/kotlin/org/wfanet/panelmatch/client/eventpreprocessing/CompressByKey.kt
@@ -22,7 +22,9 @@ import org.apache.beam.sdk.values.PCollection
 import org.apache.beam.sdk.values.PCollectionView
 import org.wfanet.panelmatch.client.combinedEvents
 import org.wfanet.panelmatch.client.common.CompressedEvents
-import org.wfanet.panelmatch.client.eventpreprocessing.EventCompressorTrainer.TrainedEventCompressor
+import org.wfanet.panelmatch.client.common.EventCompressorTrainer
+import org.wfanet.panelmatch.client.common.EventCompressorTrainer.TrainedEventCompressor
+import org.wfanet.panelmatch.client.common.TrainedEventCompressorCoder
 import org.wfanet.panelmatch.common.beam.groupByKey
 import org.wfanet.panelmatch.common.beam.kvOf
 import org.wfanet.panelmatch.common.beam.map
