+++ src/main/kotlin/org/wfanet/panelmatch/client/eventpreprocessing/PreprocessEventsInPipeline.kt
@@ -18,6 +18,7 @@ import com.google.protobuf.ByteString
 import org.apache.beam.sdk.values.KV
 import org.apache.beam.sdk.values.PCollection
 import org.wfanet.panelmatch.client.common.CompressedEvents
+import org.wfanet.panelmatch.client.common.EventCompressorTrainer
 import org.wfanet.panelmatch.common.beam.parDo
 
 /**
