+++ src/main/kotlin/org/wfanet/panelmatch/client/eventpreprocessing/deploy/gcloud/PreprocessEventsMain.kt
@@ -33,7 +33,7 @@ import org.apache.beam.sdk.options.PipelineOptionsFactory
 import org.apache.beam.sdk.options.Validation
 import org.apache.beam.sdk.values.KV
 import org.apache.beam.sdk.values.PCollection
-import org.wfanet.panelmatch.client.eventpreprocessing.BrotliEventCompressorTrainer
+import org.wfanet.panelmatch.client.common.BrotliEventCompressorTrainer
 import org.wfanet.panelmatch.client.eventpreprocessing.HardCodedDeterministicCommutativeCipherKeyProvider
 import org.wfanet.panelmatch.client.eventpreprocessing.HardCodedHkdfPepperProvider
 import org.wfanet.panelmatch.client.eventpreprocessing.HardCodedIdentifierHashPepperProvider
